+++ cpp/include/ray/api/util/function_helper.h
