+++ cpp/src/ray/util/function_helper.cc
@@ -1,5 +1,5 @@
 
-#include "function_helper.h"
+#include <ray/api/util/function_helper.h>
 
 #include <dlfcn.h>
 #include <stdio.h>
